PSR-12: refactor Aleph driver to one-class-per-file.
